Split config into several seperate files
