+++ module/utils.mjs
@@ -136,80 +136,8 @@ export function registerHandlebarsHelpers() {
 }
 
 /* -------------------------------------------- */
-/*  Config Pre-Localization                     */
+/*  Compendium Content                          */
 /* -------------------------------------------- */
-
-/**
- * Storage for pre-localization configuration.
- * @type {object}
- * @private
- */
-const _preLocalizationRegistrations = {};
-
-/**
- * Mark the provided config key to be pre-localized during the init stage.
- * @param {string} configKey              Key within `CONFIG.DND5E` to localize.
- * @param {object} [options={}]
- * @param {string} [options.key]          If each entry in the config enum is an object,
- *                                        localize and sort using this property.
- * @param {string[]} [options.keys=[]]    Array of localization keys. First key listed will be used for sorting
- *                                        if multiple are provided.
- * @param {boolean} [options.sort=false]  Sort this config enum, using the key if set.
- */
-export function preLocalize(configKey, { key, keys=[], sort=false }={}) {
-  if ( key ) keys.unshift(key);
-  _preLocalizationRegistrations[configKey] = { keys, sort };
-}
-
-/* -------------------------------------------- */
-
-/**
- * Execute previously defined pre-localization tasks on the provided config object.
- * @param {object} config  The `CONFIG.DND5E` object to localize and sort. *Will be mutated.*
- */
-export function performPreLocalization(config) {
-  for ( const [key, settings] of Object.entries(_preLocalizationRegistrations) ) {
-    _localizeObject(config[key], settings.keys);
-    if ( settings.sort ) config[key] = sortObjectEntries(config[key], settings.keys[0]);
-  }
-}
-
-/* -------------------------------------------- */
-
-/**
- * Localize the values of a configuration object by translating them in-place.
- * @param {object} obj       The configuration object to localize.
- * @param {string[]} [keys]  List of inner keys that should be localized if this is an object.
- * @private
- */
-function _localizeObject(obj, keys) {
-  for ( const [k, v] of Object.entries(obj) ) {
-    const type = typeof v;
-    if ( type === "string" ) {
-      obj[k] = game.i18n.localize(v);
-      continue;
-    }
-
-    if ( type !== "object" ) {
-      console.error(new Error(
-        `Pre-localized configuration values must be a string or object, ${type} found for "${k}" instead.`
-      ));
-      continue;
-    }
-    if ( !keys?.length ) {
-      console.error(new Error(
-        "Localization keys must be provided for pre-localizing when target is an object."
-      ));
-      continue;
-    }
-
-    for ( const key of keys ) {
-      if ( !v[key] ) continue;
-      v[key] = game.i18n.localize(v[key]);
-    }
-  }
-}
-
 
 /**
  * Synchronize the spells for all Actors in some collection with source data from an Item compendium pack.
@@ -244,6 +172,8 @@ export async function synchronizeActorSpells(actorPack, spellsPack) {
   await actorPack.configure({locked: true});
   SceneNavigation.displayProgressBar({label: "Synchronizing Spell Data", pct: 100});
 }
+
+/* -------------------------------------------- */
 
 /**
  * A helper function to synchronize spell data for a specific Actor.
